requirements.txt
selenium
pdfminer3
spacy_spanish_lemmatizer
pathos
es-lemmatizer
nltk
sklearn
pandas
pre-commit